+++ app/src/main/java/org/stepik/android/remote/billing/BillingRemoteDataSourceImpl.kt
@@ -11,8 +11,8 @@ import org.solovyev.android.checkout.Purchases
 import org.solovyev.android.checkout.RequestListener
 import org.solovyev.android.checkout.Sku
 import org.stepic.droid.di.qualifiers.MainScheduler
-import org.stepic.droid.util.consumeRx
-import org.stepic.droid.util.onReady
+import org.stepik.android.domain.billing.extension.consumeRx
+import org.stepik.android.domain.billing.extension.onReady
 import org.stepik.android.data.billing.source.BillingRemoteDataSource
 import org.stepik.android.domain.billing.exception.BillingNotSupportedException
 import org.stepik.android.view.injection.billing.SystemCheckout
