+++ billing/src/main/java/org/stepik/android/domain/billing/repository/BillingRepository.kt
@@ -1,11 +1,9 @@
 package org.stepik.android.domain.billing.repository
 
 import io.reactivex.Completable
-import io.reactivex.Maybe
 import io.reactivex.Single
 import org.solovyev.android.checkout.Purchase
 import org.solovyev.android.checkout.Sku
-import org.stepic.droid.util.maybeFirst
 
 interface BillingRepository {
     /**
@@ -16,9 +14,9 @@ interface BillingRepository {
     /**
      * Return sku
      */
-    fun getInventory(productType: String, sku: String): Maybe<Sku> =
-        getInventory(productType, listOf(sku))
-            .maybeFirst()
+//    fun getInventory(productType: String, sku: String): Maybe<Sku> =
+//        getInventory(productType, listOf(sku))
+//            .maybeFirst()
 
     /**
      * Return all available purchases (not consumed)
